Remove _fetch.go
